Move example code to examples directory
