+++ examples/TypeScript/package.json
@@ -1,0 +1,6 @@
+{
+  "devDependencies": {
+    "ts-node": "^10.9.2",
+    "typescript": "^5.8.3"
+  }
+}
